improve(cli): better interface for tests
This change is being made to expose a better interface for testing CLI
commands. This new interface has a few advantages:

1. it accepts a context.Context. This allows us to pass in a context
   from a test, to cancel long running commands (server, connector)
2. it accepts args, removing the need for a separate SetArgs call
3. it hides more of the cobra command details. We could change how the
   CLI is constructed and these tests should continue to work.

This commit also removes a few calls to context.Background. In a couple
places they are replaced with a context passed from the command, and in
others they are replaced with a TODO, to indicate we should accept the
context from the command.